+++ src/pipeline_telemetry/aggregator/__init__.py
@@ -1,2 +1,3 @@
 from .helper import TelemetryAggregator, TelemetrySelector
-from .mongo_aggregator import DailyAggregator, DailyMongoAggregator
+from .mongo_aggregator import DailyAggregator, DailyMongoAggregator, \
+    PartialToSingleAggregator, PartialToSingleMongoAggregator