+++ src/pipeline_telemetry/aggregator/aggregator.py
@@ -14,6 +14,7 @@ Usage
 >>> aggregated_telemetry = aggregator.aggregate(telemetry_queryset)
 
 """
+from abc import ABC
 from datetime import date, datetime, timedelta
 from typing import Iterator, Protocol, Type
 
@@ -46,7 +47,7 @@ class TelemetryStorage(Protocol):
         ...
 
 
-class DailyAggregator():
+class AbstractAggregator(ABC):
     """
     Aggregator to aggregate all SINGLE TELEMETRY objects for a single day
     into a telemetry objetc of type DAILY AGGREGATION.
@@ -56,12 +57,11 @@ class DailyAggregator():
 
     The following methods are available to run the actual aggregations
 
-    - aggregate: makes daily aggregations from start_date to end_date
-    - aggregate_yesteraday: makes daily aggregation for yesterday
+    - aggregate: Method to implememt the aggregation
     """
 
-    FROM_TELEMETRY_TYPE = st.SINGLE_TELEMETRY_TYPE
-    TO_TELEMETRY_TYPE = st.DAILY_AGGR_TELEMETRY_TYPE
+    FROM_TELEMETRY_TYPE: str
+    TO_TELEMETRY_TYPE: str
 
     __telemetry_selector: TelemetrySelector
     __target_telemetry: TelemetryModel
@@ -91,13 +91,6 @@ class DailyAggregator():
         for date_time_range in date_time_ranges:
             aggregated_telemetry = self._run_aggregation(date_time_range)
             self.__telemetry_storage.store_telemetry(aggregated_telemetry)
-
-    def aggregate_yesterday(self) -> None:
-        """Method to run the aggregation for yesterday.
-        """
-        start_date = date.today()
-        end_date = date.today() - timedelta(days=1)
-        self.aggregate(start_date=start_date, end_date=end_date)
 
     def _get_date_ranges(
             self, start_date: date, end_date: date) -> Iterator[DateTimeRange]:
@@ -172,3 +165,44 @@ class DailyAggregator():
             to_date_time=date_time_range.to_date,
             **telemetry_list_params
         )
+
+
+class DailyAggregator(AbstractAggregator):
+    """
+    Aggregator to aggregate all SINGLE TELEMETRY objects for a single day
+    into a telemetry objetc of type DAILY AGGREGATION.
+
+    When initializig the class a TelemetrySelector should provided that
+    determine which telemnetry objects are in scope
+
+    The following methods are available to run the actual aggregations
+
+    - aggregate: makes daily aggregations from start_date to end_date
+    - aggregate_yesterday: makes daily aggregation for yesterday
+    """
+
+    FROM_TELEMETRY_TYPE = st.SINGLE_TELEMETRY_TYPE
+    TO_TELEMETRY_TYPE = st.DAILY_AGGR_TELEMETRY_TYPE
+
+    def aggregate_yesterday(self) -> None:
+        """Method to run the aggregation for yesterday.
+        """
+        start_date = date.today()
+        end_date = date.today() - timedelta(days=1)
+        self.aggregate(start_date=start_date, end_date=end_date)
+
+
+class PartialToSingleAggregator(AbstractAggregator):
+    """
+    Aggregator to aggregate all PARTIAL TELEMETRY objects for a single day
+    into a telemetry objetc of type SINGLE AGGREGATION.
+
+    When initializig the class a TelemetrySelector should provided that
+    determine which telemnetry objects are in scope
+
+    The following methods are available to run the actual aggregations
+
+    - aggregate: makes daily aggregations from start_date to end_date
+    """
+    FROM_TELEMETRY_TYPE = st.PARTIAL_AGGR_TELEMETRY_TYPE
+    TO_TELEMETRY_TYPE = st.SINGLE_TELEMETRY_TYPE